slight re-org check stuff
