+++ neurodsp/rhythm/lc.py
@@ -5,7 +5,7 @@ from warnings import warn
 import numpy as np
 from scipy.signal.windows import hann
 
-from neurodsp.utils.core import check_n_cycles
+from neurodsp.utils.checks import check_n_cycles
 from neurodsp.utils.data import create_freqs, split_signal
 from neurodsp.utils.decorators import multidim
