+++ neurodsp/tests/utils/test_core.py
@@ -17,18 +17,3 @@ def test_get_avg_func():
 
     with raises(ValueError):
         get_avg_func('not_a_thing')
-
-def test_check_n_cycles():
-
-    n_cycles = check_n_cycles(3)
-    n_cycles = check_n_cycles([3, 4, 5])
-    n_cycles = check_n_cycles([3, 4, 5], 3)
-
-    with raises(ValueError):
-        check_n_cycles(-1)
-
-    with raises(ValueError):
-        check_n_cycles([-1, 1])
-
-    with raises(ValueError):
-        check_n_cycles([1, 2], 3)
